Display options: allow configuration of wake on plug behavior (1/2)
Adds a global setting to the Display section to enable or disable wake-on-plug.
Default is enabled.  When enabled, the device will wake when the charger
is connected or disconnected.  When disabled, the device's wake state
will be left alone.

PS 2: Fix line wrapping, set default value to false

PS 3: Fix line wrapping

PS 4: Fix line wrapping (properly this time)

Change-Id: I779a4aa4feeca87879c932be6e1fa032e4b4afc8

Conflicts:
	res/values/cm_strings.xml
	res/xml/display_settings.xml
	src/com/android/settings/DisplaySettings.java